+++ dashboard/apps/prototype/static/js/plotly-custom.js
